backend/requirements.txt
fastapi
uvicorn
google-genai
openai
fastapi
uvicorn
python-dotenv
mistralai
pydantic
python-docx
python-pptx
pymupdf
pillow
python-multipart
ipython
psycopg2-binary
qdrant-client
fastembed